clean gearannotator, finish setinstance parsing
